Middleware + token Autenticacion + env

+++ api/models/Tarea.php
@@ -1,6 +1,8 @@
 <?php
 namespace Models;
 
+use Exception;
+
 class Tarea
 {
     private $pdo;
@@ -9,14 +11,22 @@ class Tarea
     {
         $this->pdo = $pdo;
     }
-
-    public function deleteTask($idtarea)
+    public function deleteTarea($idtarea)
     {
-        $stmt = $this->pdo->prepare('DELETE FROM tarea WHERE idtarea = ?');
-        return $stmt->execute([$idtarea]);
-    }
+        try {
+            $stmt = $this->pdo->prepare('SELECT COUNT(*) FROM tarea WHERE idtarea = ?');
+            $stmt->execute([$idtarea]);
+            $count = $stmt->fetchColumn();
+            
+            if ($count == 0) throw new Exception('La tarea no existe.', 404);
 
-    public function updateTasks()
+            $stmt = $this->pdo->prepare('DELETE FROM tarea WHERE idtarea = ?');
+            return $stmt->execute([$idtarea]);
+        } catch (Exception $e) {
+            throw new Exception($e->getMessage(), $e->getCode());
+        }
+    }
+    public function updateTareas()
     {
         $stmt = $this->pdo->prepare('UPDATE tarea SET tarea_finalizada = 1 WHERE fecha_hora_modificacion < DATE_SUB(NOW(), INTERVAL 100 DAY)');
         return $stmt->execute();
